+++ CHANGELOG.md
@@ -48,6 +48,7 @@ these changes are prefixed with "**BREAKING**"
 
 * `delphin.mrs.util.rargname_sortkey()` to `delphin.sembase.role_priority()`
 * `delphin.mrs.components.Lnk` to `delphin.lnk.Lnk`
+* `delphin.mrs.components._LnkMixin` to `delphin.lnk.LnkMixin`
 * `delphin.mrs.components.split_pred_string()` to `delphin.predicate.split()`
 * `delphin.mrs.components.normalize_pred_string()` to
   `delphin.predicate.normalize()`